Refactor : 가격 예외 처리 메소드를 PriceException.class로 변경
