+++ src/main/java/lotto/exception/ExceptionController.java
@@ -1,26 +1,14 @@
-package lotto.controller;
+package lotto.exception;
 
 import lotto.model.BonusNumber;
 import lotto.model.Number;
-import lotto.model.Price;
 import lotto.model.WinningNumber;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import java.util.prefs.PreferenceChangeEvent;
 
 import static lotto.view.ErrorMessage.*;
 
 public class ExceptionController {
-    public static Price checkPriceException(int price) throws IllegalArgumentException{
-        try {
-            checkZeroPriceException(price);
-            return new Price(price);
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException();
-        }
-    }
 
     public static BonusNumber checkBonusNumberException(List<Integer> numbers, int bonus) throws IllegalArgumentException {
         try {
@@ -40,15 +28,6 @@ public class ExceptionController {
             return new WinningNumber(winningNumbers);
         } catch (IllegalArgumentException e) {
             throw e;
-        }
-    }
-
-    private static void checkZeroPriceException(int price) throws IllegalArgumentException{
-        try {
-            Price.checkZeroPrice(price);
-        } catch (IllegalArgumentException e) {
-            priceZeroException();
-            throw new IllegalArgumentException();
         }
     }
