+++ src/test/java/lotto/View/PriceTest.java
@@ -6,13 +6,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class PriceTest {
+    /*
     @DisplayName("잘못된 금액 생성 테스트. 에러 발생")
     @Test
     void wrongPriceTest() {
         Assertions.assertThatThrownBy(() -> new Price(-1)).isInstanceOf(IllegalArgumentException.class);
-    }
+    }*/
 
-    @DisplayName("올바른 금액 생성 테스트. 에러 발생")
+    @DisplayName("올바른 금액 생성 테스트")
     @Test
     void rightPriceTest() {
         Price price = new Price(1000);
